# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trashfolio API",
        "description": "Personal developer snippet manager API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/trash": {
            "get": {
                "tags": [
                    "Trash"
                ],
                "summary": "List trashed items",
                "description": "List all soft-deleted items (projects, folders, snippets) for the user.",
                "operationId": "e61f3dd2da231e72827d7c853220d00f",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of trashed items"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trash"
                ],
                "summary": "Empty trash",
                "description": "Permanently delete all trashed items belonging to the authenticated user.",
                "operationId": "6f9084031ccbee5adf2de1a0444239ad",
                "responses": {
                    "200": {
                        "description": "Trash emptied"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trash/{type}/{id}/restore": {
            "post": {
                "tags": [
                    "Trash"
                ],
                "summary": "Restore a trashed item",
                "description": "Restore a soft-deleted item by type and ID.",
                "operationId": "8ecd5d55ee3b75a45703c10d6ef20f03",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "project",
                                "folder",
                                "snippet"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item restored"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trash/{type}/{id}": {
            "delete": {
                "tags": [
                    "Trash"
                ],
                "summary": "Permanently delete a trashed item",
                "description": "Permanently delete a trashed item by type and ID.",
                "operationId": "5baad7f5d2921c1e5ed8780f63ef2034",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "project",
                                "folder",
                                "snippet"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item permanently deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}